Added repo network API call, to show forks of a given repo.  See http://develop.github.com/p/repo.html
Renamed some things in GithubDelegate for clarity; added forks array to
show forks for a selected repo in the UI.
Added tableViewSelectionDidChange behavior to UI to switch between repos
and show the resulting forks.

Added user ivar to SDGithubTaskManager for look ups on a user that may
not be the main user (i.e. the username field)